+++ lib/src/models/call_settings_model.dart
@@ -1,6 +1,9 @@
+import 'package:easy_prank_call/src/easy_prank_call_controller.dart';
+
 class CallSettingsModel {
   final Duration duration;
   final bool isVibrationOn;
+  final EasyPrankCallController controller;
 
-  const CallSettingsModel(this.duration, this.isVibrationOn);
+  const CallSettingsModel(this.duration, this.isVibrationOn, this.controller);
 }