+++ lib/src/screens/audio_call/components/audio_call_accepted_container.dart
@@ -1,5 +1,3 @@
-import 'package:easy_prank_call/src/easy_prank_call_controller.dart';
-import 'package:easy_prank_call/src/models/enums.dart';
 import 'package:easy_prank_call/src/utilities/size_config.dart';
 import 'package:easy_prank_call/src/widgets/custom_buttons.dart';
 import 'package:easy_prank_call/src/widgets/dial_button.dart';
@@ -8,10 +6,12 @@ import 'package:flutter/material.dart';
 
 class AudioCallAcceptedContainer extends StatelessWidget {
   final VoidCallback onPressEnd;
+  final VoidCallback? onPressDialButton;
 
   const AudioCallAcceptedContainer({
     Key? key,
     required this.onPressEnd,
+    this.onPressDialButton,
   }) : super(key: key);
 
   @override
@@ -28,37 +28,37 @@ class AudioCallAcceptedContainer extends StatelessWidget {
               iconData: Icons.mic,
               isDarkMode: isDarkMode,
               text: "Audio",
-              press: () => onTap(context),
+              press: onPressDialButton,
             ),
             DialButton(
               iconData: CupertinoIcons.volume_down,
               isDarkMode: isDarkMode,
               text: "Microphone",
-              press: () => onTap(context),
+              press: onPressDialButton,
             ),
             DialButton(
               iconData: CupertinoIcons.video_camera,
               isDarkMode: isDarkMode,
               text: "Video",
-              press: () => onTap(context),
+              press: onPressDialButton,
             ),
             DialButton(
               iconData: CupertinoIcons.text_bubble,
               isDarkMode: isDarkMode,
               text: "Message",
-              press: () => onTap(context),
+              press: onPressDialButton,
             ),
             DialButton(
               iconData: CupertinoIcons.person_add,
               isDarkMode: isDarkMode,
               text: "Add contact",
-              press: () => onTap(context),
+              press: onPressDialButton,
             ),
             DialButton(
               iconData: Icons.voicemail,
               isDarkMode: isDarkMode,
               text: "Voice mail",
-              press: () => onTap(context),
+              press: onPressDialButton,
             ),
           ],
         ),
@@ -71,13 +71,5 @@ class AudioCallAcceptedContainer extends StatelessWidget {
         ),
       ],
     );
-  }
-
-  void onTap(BuildContext context) {
-    final controller = EasyPrankCallController.of(context);
-    if (controller.onTapEvent != null) {
-      controller.onTapEvent!
-          .call(context, PrankCallEventAction.callScreenEvent);
-    }
   }
 }