+++ lib/src/screens/video_call/video_call_screen.dart
@@ -1,4 +1,3 @@
-import 'package:easy_prank_call/src/easy_prank_call_controller.dart';
 import 'package:easy_prank_call/src/models/call_settings_model.dart';
 import 'package:easy_prank_call/src/screens/video_call/components/body.dart';
 import 'package:flutter/material.dart';
@@ -11,10 +10,8 @@ class VideoCallScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final path = EasyPrankCallController.of(context).videoPath;
-
     return Scaffold(
-      body: Body(model.isVibrationOn, videoPath: path),
+      body: Body(model.isVibrationOn, model.controller),
     );
   }
 }