+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## 1.0.1
+## 2.0.0
 * Fixes close button bug on launch
 * Updates read me
 * Refactors widgets
+* Fixes call screen full view issue
+* removes custom navigator
+* fixes call type list tile for easy control
+* adds functionality of network video
+* 
 
 ## 1.0.1
 * Make video path optional